+++ src/components/homepage/components/Feature.jsx
@@ -2,11 +2,11 @@ import React, { useEffect, useState, useRef } from "react";
 import { useSelector } from "react-redux";
 import { Typography, Box } from "@mui/material";
 import Slider from "react-slick";
-import { fetchCards } from "../../utils/fetchCards";
-import CardComponent from "../common/Card";
+import { fetchCards } from "../../../utils/fetchCards";
+import CardComponent from "../../common/Card";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
-import "./css/feature.css"
+import "../css/feature.css";
 
 const Feature = ({ data }) => {
   const { baseUrl } = useSelector((state) => state.content);
