+++ src/components/homepage/components/ServicesSections.jsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from "react";
 import { Typography, Grid, Box } from "@mui/material";
 import { useSelector } from "react-redux";
-import { fetchCards } from "../../utils/fetchCards";
-import ServicesCard from "../common/ServicesCard";
-import "./css/servicesSections.css"
+import { fetchCards } from "../../../utils/fetchCards";
+import ServicesCard from "../../common/ServicesCard";
+import "../css/servicesSections.css";
 const ServicesSections = ({ data }) => {
   const { baseUrl } = useSelector((state) => state.content);
   const [serviceCards, setServiceCards] = useState([]);
